image upload functions extracted to utils
