+++ DipAdvisor/app/screens/AddLocationScreen.js
@@ -15,17 +15,14 @@ import { auth } from "../assets/firebase";
 import { addLocation } from "../utils/api";
 import { styles } from "../styles/styles.AddLocationScreen";
 import { useState } from "react";
-import * as ImagePicker from "expo-image-picker";
-import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
-import { storage } from "../assets/firebase";
 import "react-native-get-random-values";
-import { v4 as uuidv4 } from "uuid";
+import { uploadImage, pickImage } from "../utils/imageUploads";
 
 function AddLocationScreen({ navigation }) {
   const [image, setImage] = useState(null);
   const [imageURL, setImageURL] = useState("");
   const handlePost = (values) => {
-    uploadImage().then(() => {
+    uploadImage(image, setImageURL).then(() => {
       values.created_by = auth.currentUser.email;
       values.image_urls = imageURL;
       addLocation(values).then(({ location }) => {
@@ -34,49 +31,7 @@ function AddLocationScreen({ navigation }) {
       });
     });
   };
-  const pickImage = async () => {
-    let result = await ImagePicker.launchImageLibraryAsync({
-      mediaTypes: ImagePicker.MediaTypeOptions.All,
-      allowsEditing: true,
-      aspect: [4, 3],
-      quality: 1,
-    });
 
-    if (!result.canceled) {
-      setImage(result.assets[0].uri);
-    }
-  };
-
-  const uploadImage = async () => {
-    const storageRef = ref(storage, uuidv4());
-
-    const blob = await new Promise((resolve, reject) => {
-      const xhr = new XMLHttpRequest();
-      xhr.onload = function () {
-        resolve(xhr.response);
-      };
-      xhr.onerror = function (e) {
-        console.log(e);
-        reject(new TypeError("Network request failed"));
-      };
-      xhr.responseType = "blob";
-      xhr.open("GET", image, true);
-      xhr.send(null);
-    });
-
-    uploadBytes(storageRef, blob)
-      .then((snapshot) => {
-        console.log(snapshot);
-        console.log("Uploaded a blob!");
-        return snapshot;
-      })
-      .then((snapshot) => {
-        getDownloadURL(snapshot.ref).then((url) => {
-          setImageURL(url);
-          console.log("File available at", url);
-        });
-      });
-  };
   return (
     <ImageBackground
       style={styles.background}
@@ -118,7 +73,7 @@ function AddLocationScreen({ navigation }) {
 
                 <Button
                   title="Pick an image from camera roll"
-                  onPress={pickImage}
+                  onPress={() => pickImage(setImage)}
                 />
                 {image && (
                   <Image
