+++ DipAdvisor/app/utils/imageUploads.js
@@ -1,0 +1,50 @@
+import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
+import { storage } from "../assets/firebase";
+import "react-native-get-random-values";
+import { v4 as uuidv4 } from "uuid";
+import * as ImagePicker from "expo-image-picker";
+import "firebase/storage";
+
+export const uploadImage = async (image, setImageURL) => {
+  const storageRef = ref(storage, uuidv4());
+
+  const blob = await new Promise((resolve, reject) => {
+    const xhr = new XMLHttpRequest();
+    xhr.onload = function () {
+      resolve(xhr.response);
+    };
+    xhr.onerror = function (e) {
+      console.log(e);
+      reject(new TypeError("Network request failed"));
+    };
+    xhr.responseType = "blob";
+    xhr.open("GET", image, true);
+    xhr.send(null);
+  });
+
+  uploadBytes(storageRef, blob)
+    .then((snapshot) => {
+      console.log(snapshot);
+      console.log("Uploaded a blob!");
+      return snapshot;
+    })
+    .then((snapshot) => {
+      getDownloadURL(snapshot.ref).then((url) => {
+        setImageURL(url);
+        console.log("File available at", url);
+      });
+    });
+};
+
+export const pickImage = async (setImage) => {
+  let result = await ImagePicker.launchImageLibraryAsync({
+    mediaTypes: ImagePicker.MediaTypeOptions.All,
+    allowsEditing: true,
+    aspect: [4, 3],
+    quality: 1,
+  });
+
+  if (!result.canceled) {
+    setImage(result.assets[0].uri);
+  }
+};
